+++ adapters/adapter_with_battery.py
@@ -7,6 +7,11 @@ class AdapterWithBattery(Adapter):
         super().__init__(devices)
         self.devices.append(VoltageSensor(devices, 'cell', 'battery_voltage', ' (Battery Voltage)'))
 
+    def update_battery_status(self, device_data, message):
+        if 'battery_voltage' in message.raw:
+            device = self.get_device_by_alias('cell')
+            device.handle_message(device_data, message)
+
     def convert_message(self, message):
         message = super().convert_message(message)